Fix tables
Trac tables don't really have a header, but pandoc seems to require it (as
far as I can tell, I didn't dig into the non-simple table API). Always
treat the first row of a table as the header (this is correct for most,
but not all, of the Trac tables in our Wiki anyway).

Exclude table rows from Para end detection. Like lists, tables can start
anywhere, there doesn't need to be a blank line before the start of a
table.